feat: Add functionality to add a new game to the list of games
This commit adds a new method `AddGame` to the `GamesClient` class in order to add a new game to the list of games. The method receives a `GameDetails` object and performs the following tasks:
- Converts the `GameDetails` object to a `GameSummary` object
- Generates a new ID for the game
- Adds the new game to the list of games with the help of the `edit.razor` page

The method uses the `var` keyword to declare a variable of type `GameSummary` and initializes it with a new instance of the `GameSummary` class using object initializer syntax.

This functionality is necessary to allow the `edit.razor` page to add new games to the list of games.