feat: Inject GenresClient in EditGame page
This commit injects the `GenresClient` into the `EditGame` page to enable fetching genres for the game. The `GenresClient` is now injected using the `@inject` directive, similar to how the `GamesClient` is injected.

This change ensures that the `EditGame` page has access to the `GenresClient` and can retrieve the list of genres when initializing the page.