feat: Disable browser auto-launch in launchSettings.json
This commit updates the launchSettings.json file to disable the auto-launch of the browser when running the application in HTTPS mode. By setting the "launchBrowser" property to false, the browser will no longer automatically open when starting the application. This change improves the development experience by allowing developers to manually open the browser when needed.