feat: Add endpoint for creating a new game
This commit adds a new endpoint for creating a game in the Program.cs file. The endpoint is implemented using the MapPost function and accepts a CreateGameDto object as input. It creates a new GameDto object with the provided data and adds it to the list of games. The endpoint returns a 201 Created status code along with the newly created game. This enhancement improves the functionality of the application by allowing users to add new games.