feat: Add DELETE endpoint for removing a game
This commit adds a new DELETE endpoint in the Program.cs file to remove a game from the list of games. The endpoint is implemented using the MapDelete function and accepts the game ID as a parameter. It removes the game with the matching ID from the list of games. This enhancement improves the functionality of the application by allowing users to delete games from the system.