Add GameStoreContext class to manage database interactions
- Defined the GameStoreContext class inheriting from DbContext.
- Added a constructor that accepts DbContextOptions<GameStoreContext> to configure the context.

public class GameStoreContext: This defines a public class named GameStoreContext. This class is used to interact with a database in the context of your application.

(DbContextOptions<GameStoreContext> options): This is the constructor for the GameStoreContext class. It takes a parameter of type DbContextOptions<GameStoreContext>. This parameter is used to configure the database context with specific options.

: DbContext(options): This part calls the base class constructor (DbContext) with the options parameter. DbContext is the base class from Entity Framework Core that GameStoreContext inherits from. This allows GameStoreContext to use the options passed to it for configuring the connection to the database.